split database components into their respective databases
