+++ catkit/pawprint/__init__.py
@@ -1,4 +1,5 @@
 from .generator import Fingerprinter
+from .db import FingerprintDB
 from .operations import *
 
-__all__ = ['Fingerprinter']
+__all__ = ['Fingerprinter', 'FingerprintDB']
